Programa de Enviar SMS/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
frozenlist==1.7.0
idna==3.10
multidict==6.6.3
propcache==0.3.2
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.4
twilio==9.6.5
urllib3==2.5.0
yarl==1.20.1
